@@ -11,10 +11,10 @@ description: A new Flutter project.
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.1.0+1
+version: 1.0.0+1
 
 environment:
-  sdk: '>=2.6.0 <3.0.0'
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -24,27 +24,26 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  http: ^0.12.0+2
-  dio: ^3.0.9
-  provider: ^4.1.1
-  flutter_screenutil: ^1.1.0
-  jh_debug: ^0.3.10
-  ana_page_loop: ^0.1.7
-  shared_preferences: ^0.5.6
-  synchronized: ^2.2.0
-  ota_update: ^2.2.1
-  package_info: ^0.4.0+13 # app相关信息获取
-  url_launcher: ^5.4.2
-  permission_handler: ^5.0.0+hotfix.4 # 权限
-  get_it: ^4.0.1
-  fluttertoast: ^4.0.1
-  flare_flutter: ^2.0.3
-  assets_audio_player: ^1.4.2+1
-  flutter_spinkit: ^4.1.2+1
+  dio: 4.0.3
+  provider: 6.0.3
+  flutter_screenutil: 5.5.3+2
+  jh_debug: ^1.0.2
+  ana_page_loop: ^1.0.1
+  shared_preferences: 2.0.13 # 持久性存储
+  synchronized: 3.0.0
+  ota_update: 4.0.1
+  package_info: 2.0.2 # app相关信息获取
+  url_launcher: 6.0.20
+  permission_handler: 9.2.0 # 权限
+  fluttertoast: 8.0.8
+  flare_flutter: 3.0.2
+  assets_audio_player: 3.0.4+3
+  flutter_spinkit: 5.1.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  flutter_lints: 1.0.4
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -57,6 +56,7 @@ flutter:
   uses-material-design: true
   # To add assets to your application, add an assets section, like this:
   assets:
+    - asset/images/updateVersion/header/
     - asset/
     - asset/images/
     - asset/images/updateVersion/header/
@@ -64,11 +64,21 @@ flutter:
     - asset/bg/
     - asset/btn/
     - asset/audios/
+  #  - images/a_dot_burr.jpeg
+  #  - images/a_dot_ham.jpeg
+  # An image asset can refer to one or more resolution-specific "variants", see
+  # https://flutter.dev/assets-and-images/#resolution-aware.
+  # For details regarding adding assets from package dependencies, see
+  # https://flutter.dev/assets-and-images/#from-packages
+  # To add custom fonts to your application, add a fonts section here,
+  # in this "flutter" section. Each entry in this list should have a
+  # "family" key with the font family name, and a "fonts" key with a
+  # list giving the asset and other descriptors for the font. For
+  # example:
   fonts:
-    - family: Fish
+    - family: eye_icon
       fonts:
-        - asset: asset/fonts/Fish.ttf
-          style: italic
+        - asset: asset/fonts/EyeIcon.ttf
   #   - family: Trajan Pro
   #     fonts:
   #       - asset: fonts/TrajanPro.ttf
