@@ -1,10 +1,9 @@
 import 'package:provider/provider.dart';
 import 'package:provider/single_child_widget.dart';
-import 'pages/AppHomePage/provider/appHomePageStore.p.dart';
-import 'provider/themeStore.p.dart';
+import 'provider/global.p.dart';
+import 'provider/theme_store.p.dart';
 
 List<SingleChildWidget> providersConfig = [
-  ChangeNotifierProvider<ThemeStore>.value(value: ThemeStore()), // 主题颜色
-  // ChangeNotifierProvider<AppHomePageStore>(create: (_) => AppHomePageStore()),
-  ChangeNotifierProvider<AppHomePageStore>.value(value: AppHomePageStore()),
+  ChangeNotifierProvider<ThemeStore>(create: (_) => ThemeStore()),
+  ChangeNotifierProvider<GlobalStore>(create: (_) => GlobalStore()),
 ];
